+++ test/addContext.test.js
@@ -2,12 +2,40 @@ const addContext = require('../src/addContext');
 
 describe('addContext', () => {
   let testObj;
+  let activeTest;
   let test;
   let origConsoleError;
+
+  const makeHook = type => ({
+    title: `"${type}" hook`,
+    body: 'function () {\n    addContext(this, "i\'m in a before hook");\n  }',
+    async: 0,
+    sync: true,
+    timedOut: false,
+    pending: false,
+    type: 'hook',
+    parent: '#<Suite>',
+    ctx: '#<Context>',
+    file: '/test.js',
+    uuid: '9f4e292c-8668-4008-9dc4-589909f94054'
+  });
 
   beforeEach(() => {
     origConsoleError = console.error;
     console.error = function () {};
+
+    activeTest = {
+      title: 'sample test',
+      body: 'function () {\n    addContext(this, "i\'m in a test");\n    assert(true);\n  }',
+      async: 0,
+      sync: true,
+      timedOut: false,
+      pending: false,
+      type: 'test',
+      file: '/test.js',
+      parent: '#<Suite>',
+      ctx: '#<Context>'
+    };
   });
 
   afterEach(() => {
@@ -17,7 +45,10 @@ describe('addContext', () => {
   function contextTests() {
     it('as a string', () => {
       addContext(testObj, 'test context');
-      test.should.eql({ context: 'test context' });
+      test.should.eql({
+        ...test,
+        context: 'test context'
+      });
     });
 
     it('as an object', () => {
@@ -26,6 +57,7 @@ describe('addContext', () => {
         value: true
       });
       test.should.eql({
+        ...test,
         context: {
           title: 'context title',
           value: true
@@ -39,6 +71,7 @@ describe('addContext', () => {
         value: undefined
       });
       test.should.eql({
+        ...test,
         context: {
           title: 'context title',
           value: 'undefined'
@@ -51,14 +84,33 @@ describe('addContext', () => {
       addContext(testObj, 'test context 2');
       addContext(testObj, { title: 'test context 3', value: true });
       test.should.eql({
-        context: [ 'test context 1', 'test context 2', { title: 'test context 3', value: true } ]
+        ...test,
+        context: [
+          'test context 1',
+          'test context 2',
+          { title: 'test context 3', value: true }
+        ]
       });
     });
   }
 
   describe('when run inside a test', () => {
     beforeEach(() => {
-      testObj = { test: {} };
+      testObj = {
+        currentTest: undefined,
+        test: activeTest
+      };
+      test = testObj.test;
+    });
+    contextTests();
+  });
+
+  describe('when run inside a before', () => {
+    beforeEach(() => {
+      testObj = {
+        currentTest: activeTest,
+        test: makeHook('before all')
+      };
       test = testObj.test;
     });
     contextTests();
@@ -66,7 +118,32 @@ describe('addContext', () => {
 
   describe('when run inside a beforeEach', () => {
     beforeEach(() => {
-      testObj = { currentTest: {} };
+      testObj = {
+        currentTest: activeTest,
+        test: makeHook('before each')
+      };
+      test = testObj.currentTest;
+    });
+    contextTests();
+  });
+
+  describe('when run inside an after', () => {
+    beforeEach(() => {
+      testObj = {
+        currentTest: activeTest,
+        test: makeHook('after all')
+      };
+      test = testObj.test;
+    });
+    contextTests();
+  });
+
+  describe('when run inside an afterEach', () => {
+    beforeEach(() => {
+      testObj = {
+        currentTest: activeTest,
+        test: makeHook('after each')
+      };
       test = testObj.currentTest;
     });
     contextTests();
@@ -74,7 +151,11 @@ describe('addContext', () => {
 
   describe('No context is added when', () => {
     beforeEach(() => {
-      testObj = { test: {} };
+      testObj = {
+        test: {
+          title: 'sample test'
+        }
+      };
       test = testObj.test;
     });